This morning I added a new mode for HoverRace;  more split screen players.
You can now choose 2/3 or 4 players in split screen.

The "Setting"->"Properties" lets you set the controls for all 4 players and now there is the ability to set up to 8 joystick buttons instead of 4 (tested by Zooz and confirmed as working)

Also with an additional 2 joysticks to make it 4 total possible joysticks (yet to be tested).