There is an odd condition that sometimes occurs where a client will not successfully connect to other clients even though Wireshark shows a successful TCP handshake.  In other words, the ListCallBack() method is never getting the FD_CONNECT message for some reason.
The code I have added adds a timer that times out in five seconds; if the client connection has not been successfully established (if the FD_CONNECT message has not been received) in that time, the connection will be cleared and attempted to be re-established.

It seems that with testing that if the client connection is unsuccessful the first time, the second time will get it right.

This is a bit of a hack, and I wish I knew the underlying reasons why the FD_CONNECT message is never being received.