On Linux, use XDG Base Directory spec.
We're using libxdg-basedir to handle conforming to the spec, since it's
available in the Ubuntu and Debian package repos.

This splits the data and config directories, as per the spec, so instead of
using ~/.hoverrace for everything, we use (by default)
~/.local/share/HoverRace and ~/.config/HoverRace (these are reconfigurable
via environment vars; see the spec for details).

We still fall back to the ~/.hoverrace if the XDG base dirs can't be used
for whatever reason.

Note that Windows is unaffected, we still use %APPDATA% there.

Closes #168.