Fix bumper texture when firing engine.
When firing the engine on the basic craft, the bumper texture would
blink colors.  This was because of the wrong texture being assigned to
the first frame of the second (firing) sequence.

Note: This was only a problem with the source file that has been
committed.  The binary ObjFac1.dat we've been distributing has the
correct texture assignment.  Nobody's noticed this before because we
haven't needed to modify ObjFac1.dat.  :)