Handle string<->path_t conversion in Str.
The "P" in Str::PU and Str::UP no longer refers to the platform-specific
path string type; instead, it refers directly to path_t.  To complement
this, Str::PW (path_t to wide string) and Str::WP (wide string to
path_t) are now provided.

The upshot of this is that Str now can convert between any combination
of narrow string, wide string, and path_t.  While this complicates the
implementation, it simplifies usage of path_t significantly and hides
the differences between Boost::Filesystem v2 and v3.

Issue #184.