Flip the backbuffer in Display, not VideoBuffer.
VideoBuffer::Flip() now only blits the legacy buffer to the screen, it
doesn't do the actual flip.  Instead, the flip now occurs when the scene
has finished rendering, so that new rendering components can get a
chance to render to the backbuffer.

As a result, any scene that intends to use the legacy buffer must lock
it themselves.