Mark streams as wanting Pango formatting.
Util::SelFmt allows us to mark a stream as wanting our own objects to be
formatted in a specific way.

The use case is that when we're using Pango rendering, we want to let
our objects know that they're being written to a Pango markup string and
adjust their formatting accordingly.