Apply alignment to (approx) unscaled texture size.
The pre-scaling of the label texture (into screen-space) was causing the
alignment calculations to be off, since UiViewModel::GetAlignedPos()
expects the height and width to be in UI-space (i.e. 720p), not
screen-space.

We work around this by estimating what the UI-space size would be and
using that for alignment purposes.

Fixes #204.