Ignore return value of strerror_r() (for now).
Annoyingly, the return type of strerror_r() is defined differently
between the GNU and POSIX versions, making handling the return value
portably a pain.  At some point, the GNU version was marked with
__attribute__((warn_unused_result)) so we have to check it, but I'm not
actually sure if that's necessary given the semantics of strerror_r().

Silencing the warning for now by wrapping the call in an if.