Wrap obsolete sources in an ifdef.
These are sources which are not used in HoverRace 2.0 but are kept
around for quick-and-easy reference.

Currently, we support this by marking them as excluded in the VS project
files and by just leaving them out of the Makefile.am for Linux builds.
Adding the ifdef will allow us to not have to worry about excluding them
in the CMake build.