Link the engine statically.
We don't even pretend to provide ABI compatibility in the engine, so
most changes to the engine would require rebuild the executables anyway.
The executables will be larger, but that's not really an issue nowadays,
especially when compared to how large the resources will be once we bump
up the quality.